backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg[binary]==3.2.3
python-multipart==0.0.6
python-dotenv==1.0.0
